requirements.txt
pyserial
pyvisa-py
pyyaml